Fixes to handle out-of-order reviews
The previous version had a problem where ild reviews that were already posted were being posted again, presumably because it assumed the reviews would always be returned in order, so we only cached the last review we ever saw.

Now, we write every review ID to DynamoDB, and only post to slack if the ID didn't exist before we wrote it. That way we don't risk writing reviews we've already seen, and rely on Dynamo to manage that.

The downsides is a lot more write ops to dynamo which means we need more write capacity, which bumps the cost somewhat.